requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.4.1
Flask-WTF==1.2.2
Flask-Caching==2.3.0
redis==5.0.1
Flask-Migrate==4.0.7
WTForms==3.2.1
Werkzeug==3.1.3
psycopg[binary,pool]==3.2.10
python-socketio==5.11.4
python-engineio==4.10.1
requests==2.32.3
python-dotenv==1.0.1
pytz==2024.1
sportsipy==0.6.0
APScheduler==3.11.0
email-validator==2.2.0
bcrypt==4.2.1
gunicorn==23.0.0
eventlet==0.40.3
asgiref==3.8.1
Flask-Limiter==3.8.0